Change id from Long type to Integer type. It's provisory.

+++ api/functional/flow-creator/src/main/java/com/emanuelvictor/api/functional/flowcreator/domain/ports/repositories/IntermediaryAlternativeRepository.kt
@@ -13,7 +13,7 @@ interface IntermediaryAlternativeRepository {
     /**
      *
      */
-    fun findChildrenFromAlternativeId(id: Long): Stream<IntermediaryAlternative>
+    fun findChildrenFromAlternativeId(id: Int): Stream<IntermediaryAlternative>
 
     /**
      *
